Refactor ppba-switch to use usize for switch IDs
- Change internal switch ID type from u16 to usize to match ASCOM API
- Remove ~20 unnecessary `as u16` casts throughout device.rs
- Update MAX_SWITCH, SwitchId, SwitchInfo, and error types
- Fix clippy warning for unit struct default construction

This simplifies the code by eliminating type conversions between
the ASCOM API (which uses usize) and internal representation.

+++ services/ppba-switch/src/error.rs
@@ -25,10 +25,10 @@ pub enum PpbaError {
     ParseError(String),
 
     #[error("Invalid switch ID: {0}")]
-    InvalidSwitchId(u16),
+    InvalidSwitchId(usize),
 
     #[error("Switch not writable: {0}")]
-    SwitchNotWritable(u16),
+    SwitchNotWritable(usize),
 
     #[error("Invalid value: {0}")]
     InvalidValue(String),
